requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
jwt==1.3.1
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
psycopg==3.2.6
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0